+++ controllers/authController.js
@@ -1,6 +1,6 @@
 const HttpStatus = require('http-status-codes')
 const { createTkn } = require('../utils')
-const nodemailer = require("nodemailer")
+var mailer = require('../utils/sendMail')
 
 exports.login = async (req, res) => {
   try {
@@ -41,6 +41,9 @@ exports.login = async (req, res) => {
 
 exports.register = async (req, res) => {
   try {
+    mailer.sendMail(req.body.email);
+    console.log(req.body.email);
+
     const existingUser = await req.db.User.findOne({ email: req.body.email })
 
     if (existingUser) {
@@ -49,8 +52,6 @@ exports.register = async (req, res) => {
         message: 'User already exists!'
       })
     }
-    
-    sendVerifMail(req.body.email).catch(console.error);
 
     const user = await req.db.User.create(req.body)
 
@@ -71,28 +72,4 @@ exports.register = async (req, res) => {
       success: false
     })
   }
-}
-
-async function sendVerifMail(mail) {
-    let transporter = nodemailer.createTransport({
-      service: "Gmail",
-      auth: {
-        user: 'testdsc242@gmail.com',
-        pass: 'testdsc0'
-      }
-    });
-
-    let mailOption = {
-      from: 'Test Test <testdsc242@gmail.com>', // sender address
-      to: 'rusudaniel5799@gmail.com', // list of receivers
-      subject: "Hello", // Subject line
-      text: "Hello world?", // plain text body
-      html: "<b>Hello world?</b>" // html body --> to-do: de pus template-ul din templates/template1.html
-    };
-
-    transporter.sendMail(mailOption, error => {
-      if(error) {
-        return console.log(error);
-      }
-    });
 }
